Move ListenableFuture handling logic to BinderTransportSecurity
